requirements.txt
fastapi~=0.115.8
uvicorn~=0.34.0
python-dotenv~=1.0.1
openai~=1.61.1
langchain~=0.3.18
langchain-openai~=0.3.4
langchain-community~=0.3.17
chromadb~=0.6.3

pydantic~=2.10.6
requests~=2.32.3
